Move test to integration_cli
