+++ tests/unit_tests/gui/test_gui_load.py
@@ -1,7 +1,6 @@
 import argparse
 import os
 import shutil
-from argparse import ArgumentParser
 from pathlib import Path
 from unittest.mock import MagicMock, Mock, PropertyMock, patch
 
@@ -10,15 +9,8 @@ from qtpy.QtCore import Qt, QTimer
 from qtpy.QtWidgets import QComboBox, QMessageBox, QToolButton, QWidget
 
 import ert.gui
-from ert.__main__ import ert_parser
 from ert._c_wrappers.enkf import EnKFMain, ErtConfig
 from ert.gui.about_dialog import AboutDialog
-from ert.cli import (
-    ENSEMBLE_SMOOTHER_MODE,
-    ES_MDA_MODE,
-    ITERATIVE_ENSEMBLE_SMOOTHER_MODE,
-)
-from ert.cli.main import ErtCliError, run_cli
 from ert.gui.main import GUILogHandler, _setup_main_window, run_gui
 from ert.gui.simulation.run_dialog import RunDialog
 from ert.gui.tools.event_viewer import add_gui_log_handler
@@ -195,7 +187,6 @@ def test_that_the_ui_show_warnings_when_there_are_no_observations(qapp, tmp_path
 
 @pytest.mark.usefixtures("copy_poly_case")
 def test_that_the_ui_show_warnings_when_parameters_are_missing(qapp, tmp_path):
-
     with open("poly.ert", "r", encoding="utf-8") as fin:
         with open("poly-no-gen-kw.ert", "w", encoding="utf-8") as fout:
             for line in fin:
@@ -216,50 +207,6 @@ def test_that_the_ui_show_warnings_when_parameters_are_missing(qapp, tmp_path):
             assert not combo_box.model().item(i).isEnabled()
 
         assert gui.windowTitle() == "ERT - poly-no-gen-kw.ert"
-
-
-@pytest.mark.parametrize(
-    "mode",
-    [
-        pytest.param(ENSEMBLE_SMOOTHER_MODE),
-        pytest.param(ITERATIVE_ENSEMBLE_SMOOTHER_MODE),
-        pytest.param(ES_MDA_MODE),
-    ],
-)
-@pytest.mark.usefixtures("copy_poly_case")
-def test_that_the_cli_raises_exceptions_when_parameters_are_missing(mode):
-
-    with open("poly.ert", "r", encoding="utf-8") as fin:
-        with open("poly-no-gen-kw.ert", "w", encoding="utf-8") as fout:
-            for line in fin:
-                if "GEN_KW" not in line:
-                    fout.write(line)
-
-    args = Mock()
-    args.config = "poly-no-gen-kw.ert"
-    parser = ArgumentParser(prog="test_main")
-
-    ert_args = [
-        mode,
-        "poly-no-gen-kw.ert",
-        "--port-range",
-        "1024-65535",
-        "--target-case",
-    ]
-
-    testcase = "testcase" if mode is ENSEMBLE_SMOOTHER_MODE else "testcase-%d"
-    ert_args.append(testcase)
-
-    parsed = ert_parser(
-        parser,
-        ert_args,
-    )
-
-    with pytest.raises(
-        ErtCliError,
-        match=f"To run {mode}, GEN_KW, FIELD or SURFACE parameters are needed.",
-    ):
-        run_cli(parsed)
 
 
 @pytest.mark.usefixtures("copy_poly_case")
